Refactor the AbstractTestcontainer
