+++ src/test/java/com/bachar/AbstractTestcontainer.java
@@ -1,0 +1,44 @@
+package com.bachar;
+
+import org.flywaydb.core.Flyway;
+import org.junit.jupiter.api.BeforeAll;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+
+
+@Testcontainers
+public abstract class AbstractTestcontainer {
+    @Container
+    protected static PostgreSQLContainer<?> postgreSQLContainer =
+            new PostgreSQLContainer<>("postgres:latest")
+                    .withDatabaseName("bachar-dao-unit-test")
+                    .withUsername("bachar")
+                    .withPassword("password");
+    @BeforeAll
+    static void beforeAll() {
+        Flyway flyway = Flyway.configure().dataSource(
+                postgreSQLContainer.getJdbcUrl(),
+                postgreSQLContainer.getUsername(),
+                postgreSQLContainer.getPassword()).load();
+        flyway.migrate();
+    }
+
+    @DynamicPropertySource
+    private static void registerDataSourceProperties(DynamicPropertyRegistry registry) {
+        registry.add(
+                "spring.datasource.url",
+                postgreSQLContainer::getJdbcUrl
+        );
+        registry.add(
+                "spring.datasource.username",
+                postgreSQLContainer::getUsername
+        );
+        registry.add(
+                "spring.datasource.password",
+                postgreSQLContainer::getPassword
+        );
+    }
+}
