+++ src/test/java/com/bachar/TestcontainersTest.java
@@ -3,39 +3,11 @@ package com.bachar;
 import org.flywaydb.core.Flyway;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-@Testcontainers
-public class TestcontainersTest {
 
-    @Container
-    private static PostgreSQLContainer<?> postgreSQLContainer =
-            new PostgreSQLContainer<>("postgres:latest")
-                    .withDatabaseName("bachar-dao-unit-test")
-                    .withUsername("bachar")
-                    .withPassword("password");
-
-    @DynamicPropertySource
-    private static void registerDataSourceProperties(DynamicPropertyRegistry registry) {
-        registry.add(
-                "spring.datasource.url",
-                postgreSQLContainer::getJdbcUrl
-        );
-        registry.add(
-                "spring.datasource.username",
-                postgreSQLContainer::getUsername
-        );
-        registry.add(
-                "spring.datasource.password",
-                postgreSQLContainer::getPassword
-        );
-    }
+public class TestcontainersTest extends AbstractTestcontainer {
 
     @Test
     @DisplayName("Container created and running")
@@ -45,7 +17,8 @@ public class TestcontainersTest {
     }
 
     @Test
-    void canApplyDBMigrationsWithFlayway() {
+    @DisplayName("Migration via Flyway")
+    void canApplyDBMigrationsWithFlyway() {
         Flyway flyway = Flyway.configure().dataSource(
                 postgreSQLContainer.getJdbcUrl(),
                 postgreSQLContainer.getUsername(),
